Keep support files under spec/support
Usage of spec/dummy is causing flakiness problems. Also, `spec/dummy` seems to be unnecessary.

It is usually a good pattern to use when testing a Rails engine, but this is a Ruby gem and it is probably not a good idea.

I don't understand the difference between spec/dummy and spec/support, so I am trying to get rid of spec/dummy.